+++ functions/src/index.ts
@@ -1,9 +1,8 @@
 import { getApps, initializeApp } from "firebase-admin/app";
 
 import { createSetCommunityUserBlockCallable } from "./blockHandler";
-import { createFailClosedCommunityCallable } from "./callableScaffold";
-import { surfaceByFunctionName } from "./communityContract";
 import { createSetCreatorFollowCallable } from "./followHandler";
+import { createUpdateCreatorProfileCallable } from "./profileHandler";
 import { createSubmitCommunityReportCallable } from "./reportHandler";
 import { createFinalizeCommunitySoundUploadCallable } from "./soundUploadHandler";
 import { createRecordCommunityVoteCallable } from "./voteHandler";
@@ -19,6 +18,4 @@ export const finalizeCommunityWallpaperUpload = createFinalizeCommunityWallpaper
 export const recordCommunityVote = createRecordCommunityVoteCallable();
 export const setCreatorFollow = createSetCreatorFollowCallable();
 export const setCommunityUserBlock = createSetCommunityUserBlockCallable();
-export const updateCreatorProfile = createFailClosedCommunityCallable(
-  surfaceByFunctionName("updateCreatorProfile"),
-);
+export const updateCreatorProfile = createUpdateCreatorProfileCallable();